+++ lib/pleroma/activity.ex
@@ -310,7 +310,7 @@ defmodule Pleroma.Activity do
 
   def restrict_deactivated_users(query) do
     deactivated_users =
-      from(u in User.Query.build(deactivated: true), select: u.ap_id)
+      from(u in User.Query.build(%{deactivated: true}), select: u.ap_id)
       |> Repo.all()
 
     Activity.Queries.exclude_authors(query, deactivated_users)